backend/Lab_analysis/requirements.txt
transformers>=4.44.0
huggingface-hub>=0.24.0
torch>=2.2.0
Pillow>=10.0.0
opencv-python>=4.9.0
numpy>=1.26.0
requests>=2.31.0
matplotlib>=3.9.0
ipython>=8.25.0
pandas>=2.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0